[Layering]: Create Tensor library & move Random.
This is the first of a series of changes to implement the [Layering Swift APIs
Proposal](https://docs.google.com/document/d/1HO_sMhZJHxlDqw4Pjz4qva2s5yzgwxel-82yoK8O6L4/edit#).
This first change creates the new Tensor library and moves the random number
utilities over.
