+++ Sources/TensorFlow/Layers/Initialization.swift
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import Tensor
+
 public typealias ParameterInitializer<Scalar: TensorFlowScalar> = (TensorShape) -> Tensor<Scalar>
 
 /// Returns a function that creates a tensor by initializing all its values to zeros.
