+++ Tests/TensorFlowTests/XCTestManifests.swift
@@ -35,14 +35,12 @@ public func allTests() -> [XCTestCaseEntry] {
         testCase(LazyTensorTraceTests.allTests),
         testCase(LossTests.allTests),
         testCase(MathOperatorTests.allTests),
-        testCase(PRNGTests.allTests),
         testCase(RuntimeTests.allTests),
         testCase(SequentialTests.allTests),
         testCase(TensorAutoDiffTests.allTests),
         testCase(TensorGroupTests.allTests),
         testCase(TensorTests.allTests),
         testCase(TrivialModelTests.allTests),
-        testCase(UtilitiesTests.allTests),
     ]
 }
 #endif
