+++ Tests/TensorTests/XCTestManifests.swift
@@ -14,12 +14,12 @@
 
 import XCTest
 
-import TensorFlowTests
-import TensorTests
-import ExperimentalTests
-
-var tests = [XCTestCaseEntry]()
-tests += TensorFlowTests.allTests()
-tests += TensorTests.allTests()
-tests += ExperimentalTests.allTests()
-XCTMain(tests)
+#if !os(macOS)
+public func allTests() -> [XCTestCaseEntry] {
+    // Please ensure the test cases remain alphabetized.
+    return [
+        testCase(PRNGTests.allTests),
+        testCase(UtilitiesTests.allTests),
+    ]
+}
+#endif
